Rename name utility module
