+++ cheetah/utils/__init__.py
@@ -4,6 +4,7 @@ from .cloud_in_cell import cloud_in_cell_charge_deposition  # noqa: F401
 from .device import is_mps_available_and_functional  # noqa: F401
 from .elementwise_linspace import elementwise_linspace  # noqa: F401
 from .kde import kde_histogram_1d, kde_histogram_2d  # noqa: F401
+from .names import UniqueNameGenerator, merge_element_names  # noqa: F401
 from .physics import compute_relativistic_factors  # noqa: F401
 from .plot import (  # noqa: F401
     format_axis_as_percentage,
@@ -15,10 +16,6 @@ from .statistics import (  # noqa: F401
     unbiased_weighted_covariance_matrix,
     unbiased_weighted_std,
     unbiased_weighted_variance,
-)
-from .unique_name_generator import (  # noqa: F401
-    UniqueNameGenerator,
-    merge_element_names,
 )
 from .vector import squash_index_for_unavailable_dims  # noqa: F401
 from .warnings import (  # noqa: F401
